Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325350 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/org/apache/jmeter/reporters/Summariser.java
@@ -251,7 +251,7 @@ public class Summariser
         sb.append(longToSb(tmp,s.getCount(),5));
         sb.append(" in ");
         long elapsed = s.getElapsed();
-		sb.append(doubleToSb(tmp,(double)elapsed/1000.0,5,1));
+		sb.append(doubleToSb(tmp,elapsed/1000.0,5,1));
 		sb.append("s = ");
 		if (elapsed > 0)
 		{
